+++ include/structure/toruspoly.hpp
@@ -13,7 +13,7 @@ class DiscreteTorusPoly : public PolyBase<DiscreteTorus> {
 
   public:
     DiscreteTorusPoly() = default;
-    DiscreteTorusPoly(size_t N); 
+    DiscreteTorusPoly(uint32_t N); 
     DiscreteTorusPoly(const std::vector<DiscreteTorus> &coeffs);
 
     DiscreteTorusPoly(const DiscreteTorusPoly &toruspoly) = delete;
